requirements.txt
streamlit
pdf2image
pytesseract
spacy
langchain
google-generativeai
python-dotenv
Pillow
pandas